move `input` module from `chain` to `transaction`;
